Support for unphased genotypes.
All genotypes are treated as either phased, or unphased. Mixed phasing
is not supported in any clever way---just use the unphased option.
For unphased data, we still encode the density of minor alleles, but the
density is added across the two unphased columns of a diploid individual.
For an unresized matrix, this encodes the number of minor alleles (0, 1, or 2)
observed at a given site.
Collapsing unphased data is done before haplotype/genotype sorting.